Execute every tests even if one fails
